+++ app/src/main/java/com/decoutkhanqindev/dexreader/domain/repository/settings/SettingsRepository.kt
@@ -6,4 +6,6 @@ import kotlinx.coroutines.flow.Flow
 interface SettingsRepository {
   fun observeThemeMode(): Flow<ThemeMode>
   suspend fun saveThemeMode(value: ThemeMode)
+  fun observeIsOnboardingCompleted(): Flow<Boolean>
+  suspend fun saveIsOnboardingCompleted(value: Boolean)
 }