+++ app/src/main/java/com/decoutkhanqindev/dexreader/presentation/screens/category_details/components/filter/FilterBottomSheet.kt
@@ -79,7 +79,7 @@ fun FilterBottomSheet(
       modifier = Modifier
         .fillMaxWidth()
         .padding(horizontal = 4.dp)
-        .padding(bottom = 8.dp)
+        .padding(bottom = 16.dp)
     ) {
       Text(
         text = stringResource(R.string.apply),