+++ app/src/main/java/com/decoutkhanqindev/dexreader/presentation/screens/common/viewmodels/onboarding/OnboardingUiState.kt
@@ -1,0 +1,8 @@
+package com.decoutkhanqindev.dexreader.presentation.screens.common.viewmodels.onboarding
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+data class OnboardingUiState(
+  val isCompleted: Boolean? = null,
+)